Update and rename AdminService.java to AdminCharge.java
